[SPARK-2199][mllib] Dirichlet pdf computation moved to mllib/stats
